# Update Rust Float-Parsing Algorithms to use the Eisel-Lemire algorithm.