--- a/library/core/src/num/flt2dec/decoder.rs
+++ b/library/core/src/num/flt2dec/decoder.rs
@@ -1,6 +1,6 @@
 //! Decodes a floating-point value into individual parts and error ranges.
 
-use crate::num::dec2flt::rawfp::RawFloat;
+use crate::num::dec2flt::float::RawFloat;
 use crate::num::FpCategory;
 
 /// Decoded unsigned finite value, such that: