--- a/src/etc/test-float-parse/src/bin/huge-pow10.rs
+++ b/src/etc/test-float-parse/src/bin/huge-pow10.rs
@@ -1,6 +1,4 @@
-mod _common;
-
-use _common::validate;
+use test_float_parse::validate;
 
 fn main() {
     for e in 300..310 {