… library/core/src/num/dec2flt/algorithm.rs
… library/core/src/num/dec2flt/float.rs
… library/core/src/num/dec2flt/fpu.rs
… library/core/src/num/dec2flt/lemire.rs
… library/core/src/num/dec2flt/mod.rs
… library/core/src/num/dec2flt/num.rs
… library/core/src/num/dec2flt/number.rs
… library/core/src/num/dec2flt/parse.rs
… library/core/src/num/dec2flt/rawfp.rs
… library/core/src/num/dec2flt/slow.rs
… library/core/src/num/dec2flt/table.rs
… library/core/tests/num/dec2flt/float.rs
… library/core/tests/num/dec2flt/lemire.rs
… library/core/tests/num/dec2flt/mod.rs
… library/core/tests/num/dec2flt/parse.rs
… library/core/tests/num/dec2flt/rawfp.rs
… src/etc/dec2flt_table.py
… src/etc/test-float-parse/Cargo.toml
… src/etc/test-float-parse/runtests.py
… src/etc/test-float-parse/src/bin/many-digits.rs
… src/etc/test-float-parse/src/bin/rand-f64.rs
… src/etc/test-float-parse/src/bin/short-decimals.rs
… src/etc/test-float-parse/src/bin/subnorm.rs
… src/etc/test-float-parse/src/bin/tiny-pow10.rs
… src/etc/test-float-parse/src/bin/u32-small.rs
… src/etc/test-float-parse/src/bin/u64-pow2.rs
… src/etc/test-float-parse/src/lib.rs
… src/test/ui/issues/issue-31109.rs
… src/test/ui/issues/issue-31109.stderr
… src/test/ui/parser/float-literals.rs