@Alexhuszagh 2021-07-05
I've since finished everything with Miri (all tests pass), I've run the correctness tests again (everything passes), and re-run the binary sizes and performance tests.

tl;dr safe code is fast code. This is faster than fast-float-rust now for essentially all real-world data sets. See the summary below for a detailed explanation.

# Binary Sizes

opt-level|size|size(stripped)
|:-:|:-:|:-:|
0 | 412K | 308K
1 | 408K | 304K
2 | 392K | 296K
3 | 392K | 296K
s | 392K | 296K
z | 392K | 296K

The code compiled is as follows:

```rust
pub fn main() {
    let _ = "1.2345".parse::<f64>();
}
```

# Performance

**canada**


```text
=====================================================================================
|                         canada.txt (111126, 1.93 MB, f64)                         |
|===================================================================================|
|                                                                                   |
| ns/float                min       5%      25%   median      75%      95%      max |
|-----------------------------------------------------------------------------------|
| fast-float            28.98    29.25    29.40    29.68    29.98    30.48    35.16 |
| lexical               75.23    76.03    76.75    77.36    78.33    80.69    84.80 |
| from_str              26.08    26.84    26.98    27.11    27.42    27.97    33.04 |
|                                                                                   |
| Mfloat/s                min       5%      25%   median      75%      95%      max |
|-----------------------------------------------------------------------------------|
| fast-float            28.44    32.81    33.35    33.69    34.01    34.19    34.50 |
| lexical               11.79    12.39    12.77    12.93    13.03    13.15    13.29 |
| from_str              30.26    35.76    36.48    36.89    37.06    37.26    38.34 |
|                                                                                   |
| MB/s                    min       5%      25%   median      75%      95%      max |
|-----------------------------------------------------------------------------------|
| fast-float           494.98   570.94   580.40   586.29   591.91   594.98   600.36 |
| lexical              205.21   215.68   222.18   224.95   226.74   228.88   231.32 |
| from_str             526.62   622.26   634.72   641.86   644.89   648.42   667.15 |
|                                                                                   |
=====================================================================================
```

**uniform**


```text
=====================================================================================
|                           uniform (50000, 0.87 MB, f64)                           |
|===================================================================================|
|                                                                                   |
| ns/float                min       5%      25%   median      75%      95%      max |
|-----------------------------------------------------------------------------------|
| fast-float            25.93    26.77    27.08    27.71    27.86    28.67    38.26 |
| lexical               66.25    67.08    68.21    69.37    70.64    79.70   122.96 |
| from_str              27.74    28.66    29.28    29.74    30.26    31.36    38.93 |
|                                                                                   |
| Mfloat/s                min       5%      25%   median      75%      95%      max |
|-----------------------------------------------------------------------------------|
| fast-float            26.14    34.88    35.90    36.09    36.93    37.35    38.57 |
| lexical                8.13    12.62    14.16    14.42    14.66    14.91    15.09 |
| from_str              25.69    31.90    33.06    33.62    34.16    34.89    36.05 |
|                                                                                   |
| MB/s                    min       5%      25%   median      75%      95%      max |
|-----------------------------------------------------------------------------------|
| fast-float           455.51   607.89   625.54   628.82   643.60   650.94   672.04 |
| lexical              141.72   219.85   246.77   251.22   255.48   259.78   263.04 |
| from_str             447.66   555.96   576.02   585.88   595.31   607.97   628.25 |
|                                                                                   |
=====================================================================================
```


**one_over_rand32**

```text
=====================================================================================
|                       one_over_rand32 (50000, 1.00 MB, f64)                       |
|===================================================================================|
|                                                                                   |
| ns/float                min       5%      25%   median      75%      95%      max |
|-----------------------------------------------------------------------------------|
| fast-float            36.35    36.55    36.67    36.87    38.07    40.54    66.32 |
| lexical               87.84    88.12    88.45    89.47    92.33    98.05   120.17 |
| from_str              38.03    38.13    38.23    38.35    39.24    40.98    48.78 |
|                                                                                   |
| Mfloat/s                min       5%      25%   median      75%      95%      max |
|-----------------------------------------------------------------------------------|
| fast-float            15.08    24.67    26.27    27.12    27.27    27.36    27.51 |
| lexical                8.32    10.20    10.83    11.18    11.31    11.35    11.38 |
| from_str              20.50    24.44    25.49    26.08    26.15    26.23    26.29 |
|                                                                                   |
| MB/s                    min       5%      25%   median      75%      95%      max |
|-----------------------------------------------------------------------------------|
| fast-float           303.02   495.82   527.88   545.10   548.08   549.84   552.94 |
| lexical              167.24   205.08   217.70   224.64   227.22   228.06   228.78 |
| from_str             411.98   491.13   512.19   524.10   525.62   527.12   528.40 |
|                                                                                   |
=====================================================================================
```

**contrived cases**

| bench | core | fast-float |
|:-:|:-:|:-:|
| fast | 29.908ns | 23.798ns |
| disguised | 32.873ns | 21.378ns |
| moderate | 45.833ns | 38.855ns |
| halfway | 45.988ns | 40.700ns |
| large | 13.819us | 14.892us |
| denormal | 90.120ns | 66.922ns |


## Contrived Cases Data

This is the benchmark data for the contrived cases.

```rust
//! Benchmark sample floats meant to invoke certain code paths.

use std::time::Duration;
use fast_float::parse;
use criterion::{black_box, criterion_group, criterion_main, Criterion};

macro_rules! core_generator {
    ($group:ident, $name:literal, $str:ident, $t:ty) => {
        $group.bench_function($name, |bench| {
            bench.iter(|| {
                black_box($str.parse::<$t>().unwrap());
            })
        });
    };
}

macro_rules! fast_float_generator {
    ($group:ident, $name:literal, $str:ident, $t:ty) => {
        $group.bench_function($name, |bench| {
            bench.iter(|| {
                black_box(parse::<$t, _>($str).unwrap());
            })
        });
    };
}

// FLOATS

// NOTE: Rust currently doesn't handle large, denormal floats
// with more than 25 significant digits. Use the 25 significant
// digits for both large and denormal.

// Example fast-path value.
const FAST_PATH: &str = "1.2345e22";
// Example disguised fast-path value.
const DISGUISED_FAST_PATH: &str = "1.2345e30";
// Example moderate path value: clearly not halfway `1 << 53`.
const MODERATE_PATH: &str = "9007199254740992.0";
// Example exactly-halfway value `(1<<53) + 1`.
const HALFWAY: &str = "9007199254740993.0";
// Example large, near-halfway value.
const LARGE: &str = "8.988465674311580536566680e307";
// Example denormal, near-halfway value.
const DENORMAL: &str = "8.442911973260991817129021e-309";

fn core(criterion: &mut Criterion) {
    let mut group = criterion.benchmark_group("core");
    group.measurement_time(Duration::from_secs(5));
    core_generator!(group, "fast", FAST_PATH, f64);
    core_generator!(group, "disguised", DISGUISED_FAST_PATH, f64);
    core_generator!(group, "moderate", MODERATE_PATH, f64);
    core_generator!(group, "halfway", HALFWAY, f64);
    core_generator!(group, "large", LARGE, f64);
    core_generator!(group, "denormal", DENORMAL, f64);
}

fn fast_float(criterion: &mut Criterion) {
    let mut group = criterion.benchmark_group("fast-float");
    group.measurement_time(Duration::from_secs(5));
    fast_float_generator!(group, "fast", FAST_PATH, f64);
    fast_float_generator!(group, "disguised", DISGUISED_FAST_PATH, f64);
    fast_float_generator!(group, "moderate", MODERATE_PATH, f64);
    fast_float_generator!(group, "halfway", HALFWAY, f64);
    fast_float_generator!(group, "large", LARGE, f64);
    fast_float_generator!(group, "denormal", DENORMAL, f64);
}

// MAIN

criterion_group!(core_benches, core);
criterion_group!(fast_float_benches, fast_float);
criterion_main!(core_benches, fast_float_benches);
```

# Summary

Why do real-case benchmarks favor core, while the single float tests favor fast-float-rust? It might be due to the very aggressive inlining in fast-float-rust, which might make better code for single tests. However, in real-world examples, this is much less to be a pertinent factor.